+++ runtime/vm/object_service.cc
@@ -74,6 +74,20 @@ void Object::PrintJSON(JSONStream* stream, bool ref) const {
   }
 }
 
+void Object::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {
+  UNREACHABLE();
+}
+
+void Object::PrintImplementationFields(JSONStream* stream) const {
+  JSONObject jsobj(stream);
+  jsobj.AddProperty("type", "ImplementationFields");
+  JSONArray jsarr_fields(&jsobj, "fields");
+  if (!IsNull()) {
+    PrintImplementationFieldsImpl(jsarr_fields);
+  }
+}
+
 void Class::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Isolate* isolate = Isolate::Current();
   JSONObject jsobj(stream);
@@ -180,6 +194,9 @@ void Class::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Class::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {
+}
+
 void TypeParameters::PrintJSONImpl(JSONStream* stream, bool ref) const {
   // Consider making this type public if we decide to expose TypeParameters
   // through the protocol.
@@ -190,6 +207,9 @@ void TypeParameters::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("bounds", TypeArguments::Handle(bounds()));
   jsobj.AddProperty("defaults", TypeArguments::Handle(defaults()));
 }
+
+void TypeParameters::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void TypeArguments::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -240,9 +260,15 @@ void TypeArguments::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void TypeArguments::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void PatchClass::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
+
+void PatchClass::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Function::AddFunctionServiceId(const JSONObject& jsobj) const {
   Class& cls = Class::Handle(Owner());
@@ -388,9 +414,15 @@ void Function::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Function::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void FfiTrampolineData::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
+
+void FfiTrampolineData::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Field::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -457,6 +489,9 @@ void Field::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Field::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {
+}
+
 // See also Dart_ScriptGetTokenInfo.
 void Script::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -517,6 +552,9 @@ void Script::PrintJSONImpl(JSONStream* stream, bool ref) const {
     }
   }
 }
+
+void Script::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 static void PrintShowHideNamesToJSON(JSONObject* jsobj, const Namespace& ns) {
   Array& arr = Array::Handle();
@@ -666,17 +704,29 @@ void Library::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Library::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void LibraryPrefix::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
+
+void LibraryPrefix::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Namespace::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
 
+void Namespace::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void KernelProgramInfo::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
+
+void KernelProgramInfo::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Instructions::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -687,13 +737,22 @@ void Instructions::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Instructions::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void InstructionsSection::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
 
+void InstructionsSection::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void InstructionsTable::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
+
+void InstructionsTable::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void WeakSerializationReference::PrintJSONImpl(JSONStream* stream,
                                                bool ref) const {
@@ -704,6 +763,9 @@ void WeakSerializationReference::PrintJSONImpl(JSONStream* stream,
   auto& obj = Object::Handle(target());
   jsobj.AddProperty("target", obj);
 }
+
+void WeakSerializationReference::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void WeakArray::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -733,6 +795,9 @@ void WeakArray::PrintJSONImpl(JSONStream* stream, bool ref) const {
     }
   }
 }
+
+void WeakArray::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void ObjectPool::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -773,6 +838,9 @@ void ObjectPool::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void ObjectPool::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void PcDescriptors::PrintToJSONObject(JSONObject* jsobj, bool ref) const {
   AddCommonObjectProperties(jsobj, "Object", ref);
   // TODO(johnmccutchan): Generate a stable id. PcDescriptors hang off a Code
@@ -799,13 +867,22 @@ void PcDescriptors::PrintJSONImpl(JSONStream* stream, bool ref) const {
   PrintToJSONObject(&jsobj, ref);
 }
 
+void PcDescriptors::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void CodeSourceMap::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
 
+void CodeSourceMap::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void CompressedStackMaps::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
+
+void CompressedStackMaps::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void LocalVarDescriptors::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -833,9 +910,15 @@ void LocalVarDescriptors::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void LocalVarDescriptors::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void ExceptionHandlers::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
+
+void ExceptionHandlers::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void SingleTargetCache::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -849,6 +932,9 @@ void SingleTargetCache::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("_upperLimit", upper_limit());
 }
 
+void SingleTargetCache::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void UnlinkedCall::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   AddCommonObjectProperties(&jsobj, "Object", ref);
@@ -861,6 +947,9 @@ void UnlinkedCall::PrintJSONImpl(JSONStream* stream, bool ref) const {
                     Array::Handle(arguments_descriptor()));
 }
 
+void UnlinkedCall::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void MonomorphicSmiableCall::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   AddCommonObjectProperties(&jsobj, "Object", ref);
@@ -871,7 +960,15 @@ void MonomorphicSmiableCall::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void MonomorphicSmiableCall::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void CallSiteData::PrintJSONImpl(JSONStream* stream, bool ref) const {
+  UNREACHABLE();
+}
+
+void CallSiteData::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {
   UNREACHABLE();
 }
 
@@ -915,6 +1012,9 @@ void ICData::PrintToJSONArray(const JSONArray& jsarray,
     cache_entry.AddProperty("count", count);
   }
 }
+
+void ICData::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Code::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -978,6 +1078,8 @@ void Code::PrintJSONImpl(JSONStream* stream, bool ref) const {
   PrintJSONInlineIntervals(&jsobj);
 }
 
+void Code::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {}
+
 void Context::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   // TODO(turnidge): Should the user level type for Context be Context
@@ -1005,9 +1107,15 @@ void Context::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Context::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void ContextScope::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
+
+void ContextScope::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Sentinel::PrintJSONImpl(JSONStream* stream, bool ref) const {
   // Handle certain special sentinel values.
@@ -1034,6 +1142,9 @@ void Sentinel::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
 
+void Sentinel::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void MegamorphicCache::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   AddCommonObjectProperties(&jsobj, "Object", ref);
@@ -1048,6 +1159,9 @@ void MegamorphicCache::PrintJSONImpl(JSONStream* stream, bool ref) const {
                     Object::Handle(arguments_descriptor()));
 }
 
+void MegamorphicCache::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void SubtypeTestCache::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   AddCommonObjectProperties(&jsobj, "Object", ref);
@@ -1057,6 +1171,9 @@ void SubtypeTestCache::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
   jsobj.AddProperty("_cache", Array::Handle(cache()));
 }
+
+void SubtypeTestCache::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void LoadingUnit::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1072,7 +1189,14 @@ void LoadingUnit::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("_loadOutstanding", load_outstanding());
 }
 
+void LoadingUnit::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void Error::PrintJSONImpl(JSONStream* stream, bool ref) const {
+  UNREACHABLE();
+}
+
+void Error::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {
   UNREACHABLE();
 }
 
@@ -1084,6 +1208,9 @@ void ApiError::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("message", ToErrorCString());
 }
 
+void ApiError::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void LanguageError::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   AddCommonObjectProperties(&jsobj, "Error", ref);
@@ -1091,6 +1218,9 @@ void LanguageError::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddServiceId(*this);
   jsobj.AddProperty("message", ToErrorCString());
 }
+
+void LanguageError::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void UnhandledException::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1108,6 +1238,9 @@ void UnhandledException::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("stacktrace", instance);
 }
 
+void UnhandledException::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void UnwindError::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   AddCommonObjectProperties(&jsobj, "Error", ref);
@@ -1116,6 +1249,9 @@ void UnwindError::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("message", ToErrorCString());
   jsobj.AddProperty("_is_user_initiated", is_user_initiated());
 }
+
+void UnwindError::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Instance::PrintSharedInstanceJSON(JSONObject* jsobj,
                                        bool ref,
@@ -1185,7 +1321,15 @@ void Instance::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("kind", "PlainInstance");
 }
 
+void Instance::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void AbstractType::PrintJSONImpl(JSONStream* stream, bool ref) const {
+  UNREACHABLE();
+}
+
+void AbstractType::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {
   UNREACHABLE();
 }
 
@@ -1212,6 +1356,8 @@ void Type::PrintJSONImpl(JSONStream* stream, bool ref) const {
     jsobj.AddProperty("typeArguments", typeArgs);
   }
 }
+
+void Type::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {}
 
 void FunctionType::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1251,6 +1397,9 @@ void FunctionType::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void FunctionType::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void RecordType::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
@@ -1283,6 +1432,9 @@ void RecordType::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void RecordType::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void TypeRef::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
@@ -1295,6 +1447,9 @@ void TypeRef::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
   jsobj.AddProperty("targetType", AbstractType::Handle(type()));
 }
+
+void TypeRef::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void TypeParameter::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1314,7 +1469,15 @@ void TypeParameter::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("bound", upper_bound);
 }
 
+void TypeParameter::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void Number::PrintJSONImpl(JSONStream* stream, bool ref) const {
+  UNREACHABLE();
+}
+
+void Number::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {
   UNREACHABLE();
 }
 
@@ -1325,6 +1488,9 @@ void Integer::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("valueAsString", ToCString());
 }
 
+void Integer::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void Smi::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref, /*include_id=*/false);
@@ -1333,9 +1499,13 @@ void Smi::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddPropertyF("valueAsString", "%" Pd "", Value());
 }
 
+void Smi::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {}
+
 void Mint::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Integer::PrintJSONImpl(stream, ref);
 }
+
+void Mint::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {}
 
 void Double::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1343,6 +1513,9 @@ void Double::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("kind", "Double");
   jsobj.AddProperty("valueAsString", ToCString());
 }
+
+void Double::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void String::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1371,6 +1544,9 @@ void String::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddPropertyStr("valueAsString", *this, offset, count);
 }
 
+void String::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void Bool::PrintJSONImpl(JSONStream* stream, bool ref) const {
   const char* str = ToCString();
   JSONObject jsobj(stream);
@@ -1379,6 +1555,8 @@ void Bool::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddFixedServiceId("objects/bool-%s", str);
   jsobj.AddPropertyF("valueAsString", "%s", str);
 }
+
+void Bool::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {}
 
 void Array::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1409,6 +1587,9 @@ void Array::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Array::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {
+}
+
 void GrowableObjectArray::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
@@ -1437,6 +1618,9 @@ void GrowableObjectArray::PrintJSONImpl(JSONStream* stream, bool ref) const {
     }
   }
 }
+
+void GrowableObjectArray::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Map::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1475,6 +1659,8 @@ void Map::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Map::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {}
+
 void Set::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
@@ -1509,12 +1695,17 @@ void Set::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Set::PrintImplementationFieldsImpl(const JSONArray& jsarr_fields) const {}
+
 void Float32x4::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
   jsobj.AddProperty("kind", "Float32x4");
   jsobj.AddProperty("valueAsString", ToCString());
 }
+
+void Float32x4::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Int32x4::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1523,6 +1714,9 @@ void Int32x4::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("valueAsString", ToCString());
 }
 
+void Int32x4::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void Float64x2::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
@@ -1530,7 +1724,15 @@ void Float64x2::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("valueAsString", ToCString());
 }
 
+void Float64x2::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void TypedDataBase::PrintJSONImpl(JSONStream* stream, bool ref) const {
+  UNREACHABLE();
+}
+
+void TypedDataBase::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {
   UNREACHABLE();
 }
 
@@ -1564,9 +1766,15 @@ void TypedData::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void TypedData::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void TypedDataView::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Instance::PrintJSONImpl(stream, ref);
 }
+
+void TypedDataView::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void ExternalTypedData::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1598,17 +1806,29 @@ void ExternalTypedData::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void ExternalTypedData::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void Pointer::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Instance::PrintJSONImpl(stream, ref);
 }
+
+void Pointer::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void DynamicLibrary::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Instance::PrintJSONImpl(stream, ref);
 }
 
+void DynamicLibrary::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void Capability::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Instance::PrintJSONImpl(stream, ref);
 }
+
+void Capability::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void ReceivePort::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject obj(stream);
@@ -1622,17 +1842,29 @@ void ReceivePort::PrintJSONImpl(JSONStream* stream, bool ref) const {
   obj.AddProperty("allocationLocation", allocation_location_);
 }
 
+void ReceivePort::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void SendPort::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Instance::PrintJSONImpl(stream, ref);
 }
+
+void SendPort::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void TransferableTypedData::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Instance::PrintJSONImpl(stream, ref);
 }
 
+void TransferableTypedData::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void ClosureData::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Object::PrintJSONImpl(stream, ref);
 }
+
+void ClosureData::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Closure::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1652,6 +1884,9 @@ void Closure::PrintJSONImpl(JSONStream* stream, bool ref) const {
     jsobj.AddProperty("_activationBreakpoint", bpt);
   }
 }
+
+void Closure::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void Record::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1685,12 +1920,18 @@ void Record::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void Record::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void StackTrace::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
   jsobj.AddProperty("kind", "StackTrace");
   jsobj.AddProperty("valueAsString", ToCString());
 }
+
+void StackTrace::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void RegExp::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
@@ -1737,6 +1978,9 @@ void RegExp::PrintJSONImpl(JSONStream* stream, bool ref) const {
   }
 }
 
+void RegExp::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void WeakProperty::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
@@ -1751,6 +1995,9 @@ void WeakProperty::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("propertyValue", value_handle);
 }
 
+void WeakProperty::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void WeakReference::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
@@ -1763,7 +2010,15 @@ void WeakReference::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("target", target_handle);
 }
 
+void WeakReference::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void FinalizerBase::PrintJSONImpl(JSONStream* stream, bool ref) const {
+  UNREACHABLE();
+}
+
+void FinalizerBase::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {
   UNREACHABLE();
 }
 
@@ -1781,6 +2036,9 @@ void Finalizer::PrintJSONImpl(JSONStream* stream, bool ref) const {
   // Not exposing entries.
 }
 
+void Finalizer::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void NativeFinalizer::PrintJSONImpl(JSONStream* stream, bool ref) const {
   JSONObject jsobj(stream);
   PrintSharedInstanceJSON(&jsobj, ref);
@@ -1795,7 +2053,15 @@ void NativeFinalizer::PrintJSONImpl(JSONStream* stream, bool ref) const {
   // Not exposing entries.
 }
 
+void NativeFinalizer::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void FinalizerEntry::PrintJSONImpl(JSONStream* stream, bool ref) const {
+  UNREACHABLE();
+}
+
+void FinalizerEntry::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {
   UNREACHABLE();
 }
 
@@ -1812,17 +2078,29 @@ void MirrorReference::PrintJSONImpl(JSONStream* stream, bool ref) const {
   jsobj.AddProperty("mirrorReferent", referent_handle);
 }
 
+void MirrorReference::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void UserTag::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Instance::PrintJSONImpl(stream, ref);
 }
+
+void UserTag::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 void FutureOr::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Instance::PrintJSONImpl(stream, ref);
 }
 
+void FutureOr::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
+
 void SuspendState::PrintJSONImpl(JSONStream* stream, bool ref) const {
   Instance::PrintJSONImpl(stream, ref);
 }
+
+void SuspendState::PrintImplementationFieldsImpl(
+    const JSONArray& jsarr_fields) const {}
 
 #endif