+++ pkg/meta/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.9.0
 
+* Introduce `@reopen` to annotate class or mixin declarations that can safely
+  extend classes marked `base`, `final` or `interface`.
 * Introduce `@MustBeOverridden` to annotate class or mixin members which must be
   overridden in all subclasses.